Cleaning

+++ V_Terrain/Code/Noise/Extern/SivPerlinNoise.h
@@ -39,21 +39,21 @@ namespace siv
 
         std::int32_t p[512];
 
-        static double Fade(double t) noexcept
+        static float Fade(float t) noexcept
         {
             return t * t * t * (t * (t * 6 - 15) + 10);
         }
 
-        static double Lerp(double t, double a, double b) noexcept
+        static float Lerp(float t, float a, float b) noexcept
         {
             return a + t * (b - a);
         }
 
-        static double Grad(std::int32_t hash, double x, double y, double z) noexcept
+        static float Grad(std::int32_t hash, float x, float y, float z) noexcept
         {
             const std::int32_t h = hash & 15;
-            const double u = h < 8 ? x : y;
-            const double v = h < 4 ? y : h == 12 || h == 14 ? x : z;
+            const float u = h < 8 ? x : y;
+            const float v = h < 4 ? y : h == 12 || h == 14 ? x : z;
             return ((h & 1) == 0 ? u : -u) + ((h & 2) == 0 ? v : -v);
         }
 
@@ -79,17 +79,17 @@ namespace siv
             }
         }
 
-        double noise(double x) const
+        float noise(float x) const
         {
             return noise(x, 0.0, 0.0);
         }
 
-        double noise(double x, double y) const
+        float noise(float x, float y) const
         {
             return noise(x, y, 0.0);
         }
 
-        double noise(double x, double y, double z) const
+        float noise(float x, float y, float z) const
         {
             const std::int32_t X = static_cast<std::int32_t>(std::floor(x)) & 255;
             const std::int32_t Y = static_cast<std::int32_t>(std::floor(y)) & 255;
@@ -99,9 +99,9 @@ namespace siv
             y -= std::floor(y);
             z -= std::floor(z);
 
-            const double u = Fade(x);
-            const double v = Fade(y);
-            const double w = Fade(z);
+            const float u = Fade(x);
+            const float v = Fade(y);
+            const float w = Fade(z);
 
             const std::int32_t A = p[X] + Y, AA = p[A] + Z, AB = p[A + 1] + Z;
             const std::int32_t B = p[X + 1] + Y, BA = p[B] + Z, BB = p[B + 1] + Z;
@@ -116,10 +116,10 @@ namespace siv
                         Grad(p[BB + 1], x - 1, y - 1, z - 1))));
         }
 
-        double octaveNoise(double x, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
+        float octaveNoise(float x, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
         {
-            double result = 0.0;
-            double amp = 1.0;
+            float result = 0.0;
+            float amp = 1.0;
 
             for (std::int32_t i = 0; i < octaves; ++i)
             {
@@ -131,10 +131,10 @@ namespace siv
             return result;
         }
 
-        double octaveNoise(double x, double y, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
+        float octaveNoise(float x, float y, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
         {
-            double result = 0.0;
-            double amp = 1.0;
+            float result = 0.0;
+            float amp = 1.0;
 
             for (std::int32_t i = 0; i < octaves; ++i)
             {
@@ -147,10 +147,10 @@ namespace siv
             return result;
         }
 
-        double octaveNoise(double x, double y, double z, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
+        float octaveNoise(float x, float y, float z, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
         {
-            double result = 0.0;
-            double amp = 1.0;
+            float result = 0.0;
+            float amp = 1.0;
 
             for (std::int32_t i = 0; i < octaves; ++i)
             {
@@ -164,32 +164,32 @@ namespace siv
             return result;
         }
 
-        double noise0_1(double x) const
+        float noise0_1(float x) const
         {
             return noise(x) * 0.5 + 0.5;
         }
 
-        double noise0_1(double x, double y) const
+        float noise0_1(float x, float y) const
         {
             return noise(x, y) * 0.5 + 0.5;
         }
 
-        double noise0_1(double x, double y, double z) const
+        float noise0_1(float x, float y, float z) const
         {
             return noise(x, y, z) * 0.5 + 0.5;
         }
 
-        double octaveNoise0_1(double x, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
+        float octaveNoise0_1(float x, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
         {
             return octaveNoise(x, octaves, lacunarity, persistance) * 0.5 + 0.5;
         }
 
-        double octaveNoise0_1(double x, double y, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
+        float octaveNoise0_1(float x, float y, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
         {
             return octaveNoise(x, y, octaves, lacunarity, persistance) * 0.5 + 0.5;
         }
 
-        double octaveNoise0_1(double x, double y, double z, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
+        float octaveNoise0_1(float x, float y, float z, std::int32_t octaves, float lacunarity = 2.f, float persistance = 0.25f) const
         {
             return octaveNoise(x, y, z, octaves, lacunarity, persistance) * 0.5 + 0.5;
         }
